hillel - herman cr and remove crescendo
