--- src/main/deploy/pathplanner/paths/23.path
@@ -1,54 +1,0 @@
-{
-  "version": "2025.0",
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.55,
-        "y": 5.55
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 2.6193592783470043,
-        "y": 4.50004433322308
-      },
-      "isLocked": false,
-      "linkedName": "2"
-    },
-    {
-      "anchor": {
-        "x": 2.55,
-        "y": 4.3
-      },
-      "prevControl": {
-        "x": 2.3955341880603482,
-        "y": 4.496571394006921
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": "3"
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "pointTowardsZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 4.2,
-    "maxAcceleration": 2.0,
-    "maxAngularVelocity": 540.0,
-    "maxAngularAcceleration": 720.0,
-    "nominalVoltage": 12.0,
-    "unlimited": false
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 149.92522265446863
-  },
-  "reversed": false,
-  "folder": null,
-  "idealStartingState": {
-    "velocity": 0,
-    "rotation": -179.48349882467284
-  },
-  "useDefaultConstraints": true
-}
