--- src/main/deploy/pathplanner/paths/31.path
@@ -1,65 +1,0 @@
-{
-  "version": "2025.0",
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.55,
-        "y": 4.3
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.216928354094022,
-        "y": 4.8899327398203845
-      },
-      "isLocked": false,
-      "linkedName": "3"
-    },
-    {
-      "anchor": {
-        "x": 2.55,
-        "y": 6.85
-      },
-      "prevControl": {
-        "x": 2.0396110535810816,
-        "y": 6.79898139283458
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "pointTowardsZones": [
-    {
-      "fieldPosition": {
-        "x": 0.4,
-        "y": 5.5
-      },
-      "rotationOffset": 0.0,
-      "minWaypointRelativePos": 0.0,
-      "maxWaypointRelativePos": 1.0,
-      "name": "Point Towards Zone"
-    }
-  ],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 4.2,
-    "maxAcceleration": 2.0,
-    "maxAngularVelocity": 540.0,
-    "maxAngularAcceleration": 720.0,
-    "nominalVoltage": 12.0,
-    "unlimited": false
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": -153.3146360709931
-  },
-  "reversed": false,
-  "folder": null,
-  "idealStartingState": {
-    "velocity": 0,
-    "rotation": 151.25280049587667
-  },
-  "useDefaultConstraints": true
-}
