--- src/main/deploy/pathplanner/paths/M2.path
@@ -1,54 +1,0 @@
-{
-  "version": "2025.0",
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 1.3,
-        "y": 5.55
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 2.3000000000000007,
-        "y": 5.55
-      },
-      "isLocked": false,
-      "linkedName": "M"
-    },
-    {
-      "anchor": {
-        "x": 2.55,
-        "y": 5.55
-      },
-      "prevControl": {
-        "x": 1.5499999999999998,
-        "y": 5.55
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": "2"
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "pointTowardsZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 4.2,
-    "maxAcceleration": 2.0,
-    "maxAngularVelocity": 540.0,
-    "maxAngularAcceleration": 720.0,
-    "nominalVoltage": 12.0,
-    "unlimited": false
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 179.42583380292132
-  },
-  "reversed": false,
-  "folder": null,
-  "idealStartingState": {
-    "velocity": 0,
-    "rotation": 178.61964592655553
-  },
-  "useDefaultConstraints": true
-}
