+++ src/main/java/frc/robot/RobotState.java
@@ -1,9 +1,8 @@
 package frc.robot;
 
 public enum RobotState {
-	INTAKE,
-	SPEAKER,
-	PRE_SPEAKER,
+
 	SCORE_L4,
 	FEED
+
 }
